MNT: move helper funcions to utils
