+++ tiled/adapters/tiff.py
@@ -1,6 +1,4 @@
 import builtins
-import math
-import warnings
 from typing import Any, Dict, List, Optional, Tuple, cast
 
 import numpy as np
@@ -13,6 +11,7 @@ from ..utils import path_from_uri
 from .protocols import AccessPolicy
 from .resource_cache import with_resource_cache
 from .type_alliases import JSON, NDSlice
+from .utils import force_reshape
 
 
 class TiffAdapter:
@@ -135,70 +134,6 @@ class TiffAdapter:
 
         """
         return self._structure
-
-
-def sliced_shape(shp: Tuple[int, ...], slc: Optional[NDSlice] = ...) -> Tuple[int, ...]:
-    """Find the shape specification of an array after applying slicing"""
-
-    if slc is Ellipsis:
-        return shp
-    if isinstance(slc, int):
-        return shp[1:]
-    if isinstance(slc, builtins.slice):
-        start, stop, step = slc.indices(shp[0])
-        return (
-            max(0, (stop - start + (step - (1 if step > 0 else -1))) // step),
-            *shp[1:],
-        )
-    if isinstance(slc, tuple):
-        if len(slc) == 0:
-            return shp
-        else:
-            left_axis, *the_rest = slc
-            if (left_axis is Ellipsis) and (len(the_rest) < len(shp) - 1):
-                the_rest.insert(0, Ellipsis)
-            return *sliced_shape(shp[:1], left_axis), *sliced_shape(
-                shp[1:], tuple(the_rest)
-            )
-    return shp
-
-
-def force_reshape(
-    arr: np.array, shp: Tuple[int, ...], slc: Optional[NDSlice] = ...
-) -> np.array:
-    """Reshape a numpy array to match the desited shape, if possible.
-
-    Returns
-    -------
-
-    A view of the original array
-    """
-
-    old_shape = arr.shape
-    new_shape = sliced_shape(shp, slc)
-
-    if old_shape == new_shape:
-        # Nothing to do here
-        return arr
-
-    if math.prod(old_shape) == math.prod(new_shape):
-        if len(old_shape) != len(new_shape):
-            # Missing or extra singleton dimensions
-            warnings.warn(
-                f"Forcefully reshaping {old_shape} to {new_shape}",
-                category=RuntimeWarning,
-            )
-            return arr.reshape(new_shape)
-        else:
-            # Some dimensions might be swapped or completely wrong
-            # TODO: needs to be treated more carefully
-            pass
-
-    warnings.warn(
-        f"Can not reshape array of {old_shape} to match {new_shape}; proceeding without changes",
-        category=RuntimeWarning,
-    )
-    return arr
 
 
 class TiffSequenceAdapter:
